dmarc_function_proj/requirements.txt
tabulate
google-auth
google-auth-oauthlib
google-api-python-client
azure-identity
azure-keyvault-secrets